Change to `pynvml` instead of parsing `nvidia-smi`.
This is more robust and more efficient (takes ~1/10th of the time). I've removed some helper functions from `utils` that didn't really seem needed anymore and updated the scripts accordingly. The GPU objects returned by `get_gpu` can now contain a list of the processes running on them (for efficiency, this must be turned on with a flag if this is needed).

The output of the `gpu` command is a mess at this point, though; it just prints the list of GPU objects directly. This needs to be updated based on my fork of `gpustat` to use a nice, colored output.