Pretty printing is back!
I've added a function `get_gpu_string` that will output a nice, colorized string describing the GPU status. The format for this string is specified in `gpu_printing_config.py` which will be installed in ~/.config. Using the `gpu` command will print the output of this function, and if you call `str` on a list of GPUs, it will also return this colorized string (the list of GPUs returned is a subclass of `list` which overrides `__str__`).